orders/requirements.txt
celery==5.2.5
Django==2.2.13
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
kombu==5.2.4
psycopg2-binary==2.8.6
redis==4.2.2
ujson==5.1.0
drf_spectacular==0.22.0
django-allauth==0.50.0

pytest
pytest-django
model_bakery